Fix incorrectly exported CF_DIBV5 and CF_DIB
Reverts previous change that had switched to using the full struct size in calculation of amount of memory instead of the proper number included in the struct itself. It was then resulting with parts of the image being spliced and translated.

Also Fix exception when exporting CF_DIB as file. It was using the object/class versions of the structs. Must have forgot to change them after renaming those

Increment version to 0.3.1